Added new module appendIntoTable
