+++ index.html
@@ -12,10 +12,12 @@
       <script src="lib/jquery-3.3.1.min.js"></script>
       <script src="lib/bootstrap.min.js"></script>
       <script src="js/apiKeys.js"></script>
-      <script type="module" src="js/countryNames.js"></script>
-      <script type="module" src="services/weatherService.js"></script>
-      <script type="module" src="services/locationService.js"></script>
       <script src="services/maps.js"></script>
+      <script type="module" src="js/countryNames.js"></script>
+      <script type="module" src="modules/weather.js"></script>
+      <script type="module" src="modules/appendIntoTable.js"></script>
+      <script type="module" src="modules/weatherService.js"></script>
+      <script type="module" src="modules/locationService.js"></script>
       <script type="module" src="app.js"></script>
    </head>
